refactor(test): move SchedulerTest to nested test class in Scheduler
