+++ src/SchedulerSubsystem/Scheduler.java
@@ -1,5 +1,6 @@
 package SchedulerSubsystem;
 
+import Networking.Direction;
 import Networking.Messages.DestinationEvent;
 import Networking.Messages.ElevatorStateEvent;
 import Networking.Messages.MoveElevatorCommand;
@@ -7,8 +8,11 @@ import Networking.Messages.SystemMessage;
 import Networking.Receivers.DMA_Receiver;
 import Networking.Transmitters.DMA_Transmitter;
 import com.sun.jdi.InvalidTypeException;
+import org.junit.jupiter.api.*;
 
 import java.util.*;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 public class Scheduler implements Runnable {
     private final DMA_Transmitter transmitterToFloor;
@@ -39,7 +43,7 @@ public class Scheduler implements Runnable {
      * @param e Elevator state to check.
      * @return True if elevator should stop, false otherwise.
      */
-    public boolean isElevatorStopping(ElevatorStateEvent e) {
+    private boolean isElevatorStopping(ElevatorStateEvent e) {
         // Elevator should stop if (elevator.currentFloor, elevator.Direction)
         // belongs to the union of scheduler.destinationEvents and elevator.destinationEvents.
         //
@@ -132,6 +136,53 @@ public class Scheduler implements Runnable {
             } catch (InvalidTypeException e) {
                 throw new RuntimeException(e);
             }
+        }
+    }
+
+    /* Nested Test Class */
+
+    /**
+     * Nested test class for Scheduler unit tests.
+     *
+     * See README for reasoning.
+     */
+    public static class SchedulerTest {
+
+        DMA_Receiver receiver, mockFloorReceiver, mockElevatorReceiver;
+        DMA_Transmitter mockTransmitterToFloor, mockTransmitterToElevator;
+        Scheduler s;
+        @BeforeEach
+        public void setUp() {
+            receiver = new DMA_Receiver();
+            mockFloorReceiver = new DMA_Receiver();
+            mockElevatorReceiver = new DMA_Receiver();
+            mockTransmitterToFloor = new DMA_Transmitter(mockFloorReceiver);
+            mockTransmitterToElevator = new DMA_Transmitter((mockElevatorReceiver));
+            s = new Scheduler(receiver, mockTransmitterToFloor, mockTransmitterToElevator);
+        }
+
+        /**
+         * Positive test for isElevatorStopping.
+         */
+        @Test
+        @DisplayName("isElevatorStopping positive test")
+        public void testPositiveIsElevatorStopping() {
+            // Arrange
+            int floor = 1; // Same floor for elevator state event and destination event
+            DestinationEvent destinationEvent = new DestinationEvent(floor, Direction.UP);
+            HashSet<DestinationEvent> schedulerDestinationEvents = new HashSet<>();
+            schedulerDestinationEvents.add(destinationEvent);
+            s.setFloorRequests(schedulerDestinationEvents);
+
+            HashMap<DestinationEvent, Integer> elevatorDestinationEvents = new HashMap<DestinationEvent, Integer>();
+            elevatorDestinationEvents.put(destinationEvent, 1);
+            ElevatorStateEvent elevatorStateEvent = new ElevatorStateEvent(1, floor, Direction.UP, elevatorDestinationEvents);
+
+            // Act
+            boolean result = s.isElevatorStopping(elevatorStateEvent);
+
+            // Assert
+            assertEquals(true, result);
         }
     }
 }
